+++ services/frontend/src/app/layout/administration/administration.module.ts
@@ -1,6 +1,8 @@
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 import { SharedModule } from "src/app/core/shared/shared.module";
+import { LeftSidebarComponent } from "../header/header-siderbar/left-sidebar/left-sidebar.component";
+import { TopHeaderComponent } from "../header/header-siderbar/top-header/top-header.component";
 import { AdministrationPanelComponent } from "./administration-panel/administration-panel.component";
 import { OrganizationsComponent } from "./administration-panel/organizations/organizations.component";
 import { SuperAdminComponent } from "./administration-panel/super-admin/super-admin.component";
@@ -16,6 +18,12 @@ import { administrationRouter } from "./administration.router";
         UsersComponent,
         SuperAdminComponent,
     ],
-    imports: [CommonModule, SharedModule, administrationRouter],
+    imports: [
+        CommonModule,
+        SharedModule,
+        administrationRouter,
+        LeftSidebarComponent,
+        TopHeaderComponent,
+    ],
 })
 export class AdministrationModule {}